first half of tests up
